+++ src/main/java/command/SetReminderCommand.java
@@ -2,6 +2,7 @@ package command;
 
 import dictionary.Bank;
 import exception.ReminderSetupException;
+import parser.Parser;
 import reminder.Reminder;
 import storage.Storage;
 import ui.Ui;
@@ -45,19 +46,19 @@ public class SetReminderCommand extends Command {
     public String execute(Ui ui, Bank bank, Storage storage) {
         try {
             switch (reminderSetupState) {
-                case 1: //ask for a list of words
-                    return ui.showReminderSetup(ASK_FOR_WORDS);
-                case 2: //take in the words
-                    reminderWordList.add(userResponse);
-                    return ui.showReminderSetup(ASK_FOR_NEW_WORD);
-                case 3: //ask for reminder date
-                    return ui.showReminderSetup(ASK_FOR_REMINDER_DATE);
-                case 4:
-                    Date date = Reminder.parseDate(userResponse);
-                    new Reminder(date);
-                    return ui.showReminderSummary(reminderWordList, date);
-                default:
-                    throw new ReminderSetupException();
+            case 1: //ask for a list of words
+                return ui.showReminderSetup(ASK_FOR_WORDS);
+            case 2: //take in the words
+                reminderWordList.add(userResponse);
+                return ui.showReminderSetup(ASK_FOR_NEW_WORD);
+            case 3: //ask for reminder date
+                return ui.showReminderSetup(ASK_FOR_REMINDER_DATE);
+            case 4:
+                Date date = Parser.parseDate(userResponse);
+                new Reminder(date);
+                return ui.showReminderSummary(reminderWordList, date);
+            default:
+                throw new ReminderSetupException();
             }
         } catch (ReminderSetupException e) {
             return e.showError();
